Move all text related types into prim_store::text_run module.
